refactor: TypeScript-ify
